Remove "Queueing other entries"
This PR removes the section which could cause confusion.

Now, an entry that does not come from event dispatch logic does not need to have processingStart and processingEnd set to 0. And we no longer force the first input to go through event dispatch logic as that seems non-trivial performance-wise. Thus, removing the section and adding a small note.