Fix description of processingStart/End in spec
The non-normative section describing the processingStart and processingEnd attributes allows them to be set to 0, but we've decided to not allow this for now. Thus, we remove this from the description. The non-normative note describing how they are set when the event dispatching logic is skipped (by 'estimating' them) is not removed.