Set EventCounts map
This is perhaps a bit hand-wavy, but I notice hr-time doesn't even mention the construction of the Performance object so...

One question is whether the user agent should set all the possible events to 0, or only those which it can count. This is relevant for Chromium implementation as 'dragexit' is not implemented (in Chrome or elsewhere, but it is in HTML... that might change after https://github.com/whatwg/html/issues/2741 is resolved). In any case, it seems better to initialize to 0s only for event types you're currently counting.