Remove pointermove, interactionType, and pointerIsDragSet from the interactionId computation
Fixes #140.

`pointermove` is not exposed as an event entry, so it doesn't make sense
to check it. `interactionType` is only set in the spec, but not used.
And `pointerIsDragSet` was only used for the aforementioned
`interactionType`. It would be nice to clean up these and simplify the
`interactionId` computation logic.