Reset the "is contextmenu triggered" flag on pointerup
Given the current spec wording, it's possible to have a pointerup
without a pointerdown event to pair with. Normally this case shouldn't
happen, but it looks like Chromium experienced this in the field in very
early stages of page load, before the page had a chance to bootstrap
event listeners.

This commit makes sure that we reset this flag state on pointerup as
well, effectively making it a single-use flag that resets either on
pointerup or on the next pointerdown.