Add `contextmenu` to interactionId computation and handle the right click cases (#156)
* Add `contextmenu` to interactionId computation and handle the right click cases

This patch does 2 things:

1. Adds interactionId to the `contextmenu` events.
2. Makes sure that all the `pointerdown` events have known
   interactionIds.

The 2nd item was an issue we encountered in Firefox. Currently Chrome
uses a timeout mechanism to handle the right clicks, but I believe this
solution handles all the cases in all platforms, see
https://gist.github.com/canova/c7c87530e68003c4ac9d8e936cf80664.

* Reset the "is contextmenu triggered" flag on pointerup

Given the current spec wording, it's possible to have a pointerup
without a pointerdown event to pair with. Normally this case shouldn't
happen, but it looks like Chromium experienced this in the field in very
early stages of page load, before the page had a chance to bootstrap
event listeners.

This commit makes sure that we reset this flag state on pointerup as
well, effectively making it a single-use flag that resets either on
pointerup or on the next pointerdown.